Now generating enums and partially handling functions.
For functions to work we need to split out inlined functions into a separate class (which we'll use to load against the bridgesupport dylibs.)

To get this to work, we're now generating "opaque" types from the XML as subclasses of JNA's PointerType.

TODO:
Handling inlines
Clojure wrappers for the generated functions
Constants (this is a big-ish job as we'll need to invoke dlsym ourselves, and it would help to have the lib handle.)
Clean up the spaghetti in bstool.clj

KNOWN BUGS:
CoreSupport breaks us for some reason.